refactor(composition): relocate the profile edge into deployment.rs; branching ratchet 3->2 (#6274)
`local_runtime_profile.rs` was a genuine composition edge — the CLI-facing
profile->{config, policy, build-input} mapping (`deployment_config_for_profile`,
`local_runtime_build_input*`, the `*_runtime_policy` accessors) — with five
production sites naming a `RebornCompositionProfile`/`RuntimeProfile` variant.
That is exactly what `deployment.rs::for_profile` already is, so this makes
`deployment.rs` the single place a profile name becomes deployment data (§4.4).

- Move the error/options types and all profile->{config,policy,build-input}
  functions from `local_runtime_profile.rs` into `deployment.rs`; move its tests
  as `#[cfg(test)] mod local_runtime_profile_tests`. Delete the module.
- lib.rs: drop `mod local_runtime_profile;`; the `pub use` facade now re-exports
  the same symbols from `deployment` — every caller uses the crate-level
  re-export, so no call site changes.
- Ratchet allowlist 3 -> 2: `{deployment.rs (TARGET), readiness.rs}`.
  `readiness.rs` remains (profile is an operator-facing wire label, not a
  branch).
- Regenerated the composition pub-use snapshot (re-export path only).

Pure relocation — no behavior change. The hosted-single-tenant-volume build
path (incl. its `libsql`-feature fail-closed guard) moved verbatim; the
`profile_acceptance` volume readiness + `MissingLibsqlFeature` guards still pass.

[skip-regression-check] module relocation + ratchet-allowlist trim; the
architecture ratchet + the moved profile_acceptance tests are the pinning tests.
